added model with react
