+++ Model_Demo/zewit/src/index.css
@@ -11,3 +11,11 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+#root{
+  margin: 0% !important;
+  padding:0% !important;
+}
+html,body,.app{
+  margin: 0% !important;
+  padding:0% !important;
+}
